Adding a game to our DB from BGG call
When looking up a game by bgg ID, first looks in the DB to see if we've
already added it, if so, return it.  Else go to the WS, get it, store it
in the DB, then return it

Also added a test for converting WS data into our game model.
Added a secondary grunt task to run tests when changind different levels
of our app.