Display Flag
Added a display flag attribute to the game detail output.  This will
allow the user to set a flag to decide if they would like to display
this game in a list of all games.  (e.g. maybe they don't want to show
Cards Against Humanity because no one wants to play it, so why bother
showing in a list of games to play).